[chore] MemberRepository findById 파라미터 변경

+++ src/main/java/com/example/springtrash/member/repository/MemberRepositoryImpl.java
@@ -25,7 +25,7 @@ public class MemberRepositoryImpl implements MemberRepository {
     }
 
     @Override
-    public Optional<Member> findById(Integer id) {
+    public Optional<Member> findById(Long id) {
         return Optional.empty();
     }
